# Minimize substitutions inside inlineCleanup