before:
```
benchmarking normalization of examples/Reducer.hs
time                 2.280 s    (2.156 s .. 2.463 s)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 2.268 s    (2.242 s .. 2.283 s)
std dev              25.00 ms   (6.513 ms .. 33.26 ms)
variance introduced by outliers: 19% (moderately inflated)
```

after:
```
benchmarking normalization of examples/Reducer.hs
time                 1.002 s    (969.5 ms .. 1.025 s)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 1.019 s    (1.007 s .. 1.023 s)
std dev              8.327 ms   (49.15 μs .. 10.22 ms)
variance introduced by outliers: 19% (moderately inflated)
```